+++ vscode/src/chat/protocol.ts
@@ -199,6 +199,7 @@ export type ExtensionMessage =
           configFeatures: {
               chat: boolean
               attribution: boolean
+              serverSentModels: boolean
           }
       }
     | {