+++ vscode/webviews/App.tsx
@@ -9,11 +9,10 @@ import {
     type CodyCommand,
     CodyIDE,
     GuardrailsPost,
-    Model,
+    type Model,
     PromptString,
     type SerializedChatTranscript,
     isCodyProUser,
-    isEnterpriseUser,
 } from '@sourcegraph/cody-shared'
 import type { AuthMethod, ConfigurationSubsetForWebview, LocalEnv } from '../src/chat/protocol'
 import type { UserAccountInfo } from './Chat'
@@ -53,6 +52,7 @@ export const App: React.FunctionComponent<{ vscodeAPI: VSCodeWrapper }> = ({ vsc
     const [chatEnabled, setChatEnabled] = useState<boolean>(true)
     const [attributionEnabled, setAttributionEnabled] = useState<boolean>(false)
     const [commandList, setCommandList] = useState<CodyCommand[]>([])
+    const [serverSentModelsEnabled, setServerSentModelsEnabled] = useState<boolean>(false)
 
     const [clientState, setClientState] = useState<ClientStateForWebview>({
         initialContext: [],
@@ -106,8 +106,6 @@ export const App: React.FunctionComponent<{ vscodeAPI: VSCodeWrapper }> = ({ vsc
                             // Receive this value from the extension backend to make it work
                             // with E2E tests where change the DOTCOM_URL via the env variable TESTING_DOTCOM_URL.
                             isDotComUser: message.authStatus.isDotCom,
-                            // Default to assuming they are a single model enterprise
-                            isOldStyleEnterpriseUser: isEnterpriseUser(message.authStatus),
                             user: message.authStatus,
                             ide: message.config.agentIDE || CodyIDE.VSCode,
                         })
@@ -121,6 +119,7 @@ export const App: React.FunctionComponent<{ vscodeAPI: VSCodeWrapper }> = ({ vsc
                     case 'setConfigFeatures':
                         setChatEnabled(message.configFeatures.chat)
                         setAttributionEnabled(message.configFeatures.attribution)
+                        setServerSentModelsEnabled(message.configFeatures.serverSentModels)
                         break
                     case 'history':
                         setUserHistory(Object.values(message.localHistory?.chat ?? {}))
@@ -145,15 +144,6 @@ export const App: React.FunctionComponent<{ vscodeAPI: VSCodeWrapper }> = ({ vsc
                         break
                     case 'chatModels':
                         setChatModels(message.models)
-                        setUserAccountInfo(
-                            info =>
-                                info && {
-                                    ...info,
-                                    isOldStyleEnterpriseUser:
-                                        !info.isDotComUser &&
-                                        !message.models.some(Model.isNewStyleEnterprise),
-                                }
-                        )
                         break
                     case 'attribution':
                         if (message.attribution) {
@@ -175,6 +165,19 @@ export const App: React.FunctionComponent<{ vscodeAPI: VSCodeWrapper }> = ({ vsc
             }),
         [errorMessages, view, vscodeAPI, guardrails, dispatchClientAction]
     )
+
+    useEffect(() => {
+        // On macOS, suppress the '¬' character emitted by default for alt+L
+        const handleKeyDown = (event: KeyboardEvent) => {
+            if (event.altKey && event.key === '¬') {
+                event.preventDefault()
+            }
+        }
+        window.addEventListener('keydown', handleKeyDown)
+        return () => {
+            window.removeEventListener('keydown', handleKeyDown)
+        }
+    }, [])
 
     useEffect(() => {
         // Notify the extension host that we are ready to receive events
@@ -228,8 +231,8 @@ export const App: React.FunctionComponent<{ vscodeAPI: VSCodeWrapper }> = ({ vsc
         [vscodeAPI]
     )
     const chatModelContext = useMemo<ChatModelContext>(
-        () => ({ chatModels, onCurrentChatModelChange }),
-        [chatModels, onCurrentChatModelChange]
+        () => ({ chatModels, onCurrentChatModelChange, serverSentModelsEnabled }),
+        [chatModels, onCurrentChatModelChange, serverSentModelsEnabled]
     )
 
     // Wait for all the data to be loaded before rendering Chat View