+++ web/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.7
+- Provide api to set custom headers for any network request in Cody Web (agent worker thread)
+
+## 0.2.6
+- Don't show model-select dropdown for enterprise users who don't have server-sent models enabled
+
 ## 0.2.5
 - Don't send vscode-related init telemetry events 
 - Fix enterprise LLM selector appearance checks for Cody Web